Refactor: reorganize LoadBalancer controller by functionality
Split controller.go (543 lines) into focused files:
- controller.go (215 lines): Core reconciliation logic
- instance.go (72 lines): NFQLB instance management
- targets.go (116 lines): Target activation/deactivation
- flows.go (235 lines): Flow configuration from L34Routes

Benefits:
- Improved code organization and readability
- Easier to maintain and test individual components
- Clear separation of concerns
- All 24 tests passing
