+++ internal/controller/loadbalancer/controller.go
@@ -18,12 +18,9 @@ package loadbalancer
 
 import (
 	"context"
-	"fmt"
-	"strconv"
 	"sync"
 
 	"github.com/google/nftables"
-	nspAPI "github.com/nordix/meridio/api/nsp/v1"
 	"github.com/nordix/meridio/pkg/loadbalancer/types"
 	discoveryv1 "k8s.io/api/discovery/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -70,10 +67,7 @@ type Controller struct {
 	flows     map[string]map[string]*meridio2v1alpha1.L34Route // key: DistributionGroup name -> L34Route name
 }
 
-const defaultMaxEndpoints = 32
 const kindDistributionGroup = "DistributionGroup"
-
-const identifierOffset = 5000 // TODO: port identifierOffsetGenerator from Meridio
 
 func (c *Controller) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	logr := log.FromContext(ctx)
@@ -141,16 +135,7 @@ func (c *Controller) belongsToGateway(ctx context.Context, distGroup *meridio2v1
 		route := &l34routeList.Items[i]
 
 		// Check if route references this Gateway
-		referencesGateway := false
-		for _, parentRef := range route.Spec.ParentRefs {
-			if string(parentRef.Name) == c.GatewayName &&
-				(parentRef.Namespace == nil || string(*parentRef.Namespace) == c.GatewayNamespace) {
-				referencesGateway = true
-				break
-			}
-		}
-
-		if !referencesGateway {
+		if !c.referencesGateway(route) {
 			continue
 		}
 
@@ -165,314 +150,6 @@ func (c *Controller) belongsToGateway(ctx context.Context, distGroup *meridio2v1
 	}
 
 	return false
-}
-
-func (c *Controller) reconcileNFQLBInstance(ctx context.Context, distGroup *meridio2v1alpha1.DistributionGroup) error {
-	logr := log.FromContext(ctx)
-
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	// Initialize maps if needed
-	if c.instances == nil {
-		c.instances = make(map[string]types.NFQueueLoadBalancer)
-	}
-	if c.targets == nil {
-		c.targets = make(map[string]map[int][]string)
-	}
-
-	// Check if instance already exists
-	if _, exists := c.instances[distGroup.Name]; exists {
-		return nil
-	}
-
-	// Get Maglev parameters
-	n := int32(defaultMaxEndpoints)
-	if distGroup.Spec.Maglev != nil {
-		n = distGroup.Spec.Maglev.MaxEndpoints
-	}
-	m := int(n * 100) // M = N × 100
-
-	// Create NFQLB instance
-	instance, err := c.LBFactory.New(distGroup.Name, m, int(n))
-	if err != nil {
-		return err
-	}
-
-	// Start the instance to create shared memory
-	if err := instance.Start(); err != nil {
-		logr.Error(err, fmt.Sprintf("failed to start NFQLB instance, distGroup: %s", distGroup.Name))
-		return err
-	}
-
-	c.instances[distGroup.Name] = instance
-	c.targets[distGroup.Name] = make(map[int][]string)
-
-	logr.Info(fmt.Sprintf("Created NFQLB instance, distGroup: %s, M: %d, N: %d", distGroup.Name, m, n))
-	return nil
-}
-
-func (c *Controller) reconcileTargets(ctx context.Context, distGroup *meridio2v1alpha1.DistributionGroup) error {
-	logr := log.FromContext(ctx)
-
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	instance, exists := c.instances[distGroup.Name]
-	if !exists {
-		return nil
-	}
-
-	// Get EndpointSlices for this DistributionGroup
-	endpointSliceList := &discoveryv1.EndpointSliceList{}
-	if err := c.List(ctx, endpointSliceList,
-		client.InNamespace(c.GatewayNamespace),
-		client.MatchingLabels{
-			"meridio-2.nordix.org/distributiongroup": distGroup.Name,
-		}); err != nil {
-		return err
-	}
-
-	// Get current targets
-	currentTargets := c.targets[distGroup.Name]
-	if currentTargets == nil {
-		currentTargets = make(map[int][]string)
-		c.targets[distGroup.Name] = currentTargets
-	}
-
-	// Build new targets map from EndpointSlices
-	newTargets := make(map[int][]string)
-	for _, eps := range endpointSliceList.Items {
-		for _, endpoint := range eps.Endpoints {
-			if endpoint.Conditions.Ready == nil || !*endpoint.Conditions.Ready {
-				continue
-			}
-			if endpoint.Zone == nil {
-				logr.V(1).Info("Endpoint missing identifier (Zone field)", "addresses", endpoint.Addresses)
-				continue
-			}
-
-			// Parse Zone field - support both "maglev:N" and plain "N" formats
-			zoneStr := *endpoint.Zone
-			if len(zoneStr) > 7 && zoneStr[:7] == "maglev:" {
-				zoneStr = zoneStr[7:] // Strip "maglev:" prefix
-			}
-
-			identifier, err := strconv.Atoi(zoneStr)
-			if err != nil {
-				logr.Error(err, "Invalid identifier in Zone field", "zone", *endpoint.Zone)
-				continue
-			}
-			newTargets[identifier] = endpoint.Addresses
-		}
-	}
-
-	// Deactivate removed targets
-	for identifier := range currentTargets {
-		if _, exists := newTargets[identifier]; !exists {
-			if err := instance.Deactivate(identifier + 1); err != nil {
-				logr.Error(err, "Failed to deactivate target", "identifier", identifier)
-			} else {
-				logr.Info("Deactivated target", "distGroup", distGroup.Name, "identifier", identifier)
-			}
-		}
-	}
-
-	// Activate new targets
-	for identifier, ips := range newTargets {
-		if _, exists := currentTargets[identifier]; !exists {
-			// NFQLB expects 1-based index, so add 1 to 0-based identifier
-			// Fwmark is identifier + offset for routing
-			if err := instance.Activate(identifier+1, identifier+identifierOffset); err != nil {
-				logr.Error(err, "Failed to activate target", "identifier", identifier, "ips", ips)
-			} else {
-				logr.Info("Activated target", "distGroup", distGroup.Name, "identifier", identifier, "ips", ips)
-			}
-		}
-	}
-
-	// Update tracked targets
-	c.targets[distGroup.Name] = newTargets
-
-	logr.Info("Reconciled targets", "distGroup", distGroup.Name, "count", len(newTargets))
-	return nil
-}
-
-func (c *Controller) reconcileFlows(ctx context.Context, distGroup *meridio2v1alpha1.DistributionGroup) error {
-	logr := log.FromContext(ctx)
-
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	instance, exists := c.instances[distGroup.Name]
-	if !exists {
-		return nil
-	}
-
-	// Initialize flows map if needed
-	if c.flows == nil {
-		c.flows = make(map[string]map[string]*meridio2v1alpha1.L34Route)
-	}
-	if c.flows[distGroup.Name] == nil {
-		c.flows[distGroup.Name] = make(map[string]*meridio2v1alpha1.L34Route)
-	}
-
-	// Check if DistributionGroup has endpoints before configuring flows
-	endpointSliceList := &discoveryv1.EndpointSliceList{}
-	if err := c.List(ctx, endpointSliceList,
-		client.InNamespace(c.GatewayNamespace),
-		client.MatchingLabels{
-			"meridio-2.nordix.org/distributiongroup": distGroup.Name,
-		}); err != nil {
-		return err
-	}
-
-	hasEndpoints := false
-	for _, eps := range endpointSliceList.Items {
-		for _, endpoint := range eps.Endpoints {
-			if endpoint.Conditions.Ready != nil && *endpoint.Conditions.Ready {
-				hasEndpoints = true
-				break
-			}
-		}
-		if hasEndpoints {
-			break
-		}
-	}
-
-	if !hasEndpoints {
-		logr.Info("No ready endpoints, deleting flows", "distGroup", distGroup.Name)
-		// Delete all flows for this DistributionGroup
-		currentFlows := c.flows[distGroup.Name]
-		var errFinal error
-		for flowName := range currentFlows {
-			flow := &nspAPI.Flow{Name: flowName}
-			if err := instance.DeleteFlow(flow); err != nil {
-				logr.Error(err, "Failed to delete flow", "flow", flowName)
-				errFinal = fmt.Errorf("%w; failed to delete flow %s: %w", errFinal, flowName, err)
-			} else {
-				logr.Info("Deleted flow", "distGroup", distGroup.Name, "flow", flowName)
-			}
-		}
-		c.flows[distGroup.Name] = make(map[string]*meridio2v1alpha1.L34Route)
-		return errFinal
-	}
-
-	// Get L34Routes for this Gateway and DistributionGroup
-	l34routeList := &meridio2v1alpha1.L34RouteList{}
-	if err := c.List(ctx, l34routeList, client.InNamespace(c.GatewayNamespace)); err != nil {
-		return err
-	}
-
-	// Build new flows map
-	newFlows := make(map[string]*meridio2v1alpha1.L34Route)
-	for i := range l34routeList.Items {
-		route := &l34routeList.Items[i]
-
-		// Check if route references this Gateway
-		referencesGateway := false
-		for _, parentRef := range route.Spec.ParentRefs {
-			if string(parentRef.Name) == c.GatewayName &&
-				(parentRef.Namespace == nil || string(*parentRef.Namespace) == c.GatewayNamespace) {
-				referencesGateway = true
-				break
-			}
-		}
-		if !referencesGateway {
-			continue
-		}
-
-		// Check if route references this DistributionGroup
-		referencesDistGroup := false
-		var dgName string
-		for _, backendRef := range route.Spec.BackendRefs {
-			if backendRef.Group != nil && string(*backendRef.Group) == meridio2v1alpha1.GroupVersion.Group &&
-				backendRef.Kind != nil && string(*backendRef.Kind) == "DistributionGroup" &&
-				string(backendRef.Name) == distGroup.Name {
-				referencesDistGroup = true
-				dgName = string(backendRef.Name)
-				break
-			}
-		}
-		if !referencesDistGroup {
-			continue
-		}
-
-		// Validate that referenced DistributionGroup exists
-		var dg meridio2v1alpha1.DistributionGroup
-		if err := c.Get(ctx, client.ObjectKey{Name: dgName, Namespace: c.GatewayNamespace}, &dg); err != nil {
-			if apierrors.IsNotFound(err) {
-				logr.Info("Skipping L34Route: DistributionGroup not found", "route", route.Name, "distributionGroup", dgName)
-				continue
-			}
-			return err
-		}
-
-		newFlows[route.Name] = route
-	}
-
-	// Delete removed flows
-	currentFlows := c.flows[distGroup.Name]
-	var errFinal error
-	for flowName := range currentFlows {
-		if _, exists := newFlows[flowName]; !exists {
-			flow := &nspAPI.Flow{Name: flowName}
-			if err := instance.DeleteFlow(flow); err != nil {
-				logr.Error(err, "Failed to delete flow", "flow", flowName)
-				errFinal = fmt.Errorf("%w; failed to delete flow %s: %w", errFinal, flowName, err)
-			} else {
-				logr.Info("Deleted flow", "distGroup", distGroup.Name, "flow", flowName)
-			}
-		}
-	}
-
-	// Add/update flows
-	for flowName, route := range newFlows {
-		flow := c.convertL34RouteToFlow(route)
-		if err := instance.SetFlow(flow); err != nil {
-			logr.Error(err, "Failed to set flow", "flow", flowName)
-			// TODO(P2): Detect NFQLB capacity errors and update DistributionGroup status
-			// if strings.Contains(err.Error(), "capacity exceeded") {
-			//     return fmt.Errorf("NFQLB capacity exceeded: %w", err)
-			// }
-			errFinal = fmt.Errorf("%w; failed to set flow %s: %w", errFinal, flowName, err)
-		} else {
-			logr.Info("Configured flow", "distGroup", distGroup.Name, "flow", flowName)
-		}
-	}
-
-	// Update tracked flows
-	c.flows[distGroup.Name] = newFlows
-
-	logr.Info("Reconciled flows", "distGroup", distGroup.Name, "count", len(newFlows))
-	return errFinal
-}
-
-func (c *Controller) convertL34RouteToFlow(route *meridio2v1alpha1.L34Route) *nspAPI.Flow {
-	flow := &nspAPI.Flow{
-		Name:                  route.Name,
-		Priority:              route.Spec.Priority,
-		SourceSubnets:         route.Spec.SourceCIDRs,
-		SourcePortRanges:      route.Spec.SourcePorts,
-		DestinationPortRanges: route.Spec.DestinationPorts,
-		ByteMatches:           route.Spec.ByteMatches,
-	}
-
-	// Convert protocols
-	protocols := make([]string, len(route.Spec.Protocols))
-	for i, p := range route.Spec.Protocols {
-		protocols[i] = string(p)
-	}
-	flow.Protocols = protocols
-
-	// Convert VIPs
-	vips := make([]*nspAPI.Vip, len(route.Spec.DestinationCIDRs))
-	for i, cidr := range route.Spec.DestinationCIDRs {
-		vips[i] = &nspAPI.Vip{Address: cidr}
-	}
-	flow.Vips = vips
-
-	return flow
 }
 
 func (c *Controller) SetupWithManager(mgr ctrl.Manager) error {
@@ -513,15 +190,7 @@ func (c *Controller) l34RouteEnqueue(ctx context.Context, obj client.Object) []c
 	}
 
 	// Check if route references this Gateway
-	referencesGateway := false
-	for _, parentRef := range route.Spec.ParentRefs {
-		if string(parentRef.Name) == c.GatewayName &&
-			(parentRef.Namespace == nil || string(*parentRef.Namespace) == c.GatewayNamespace) {
-			referencesGateway = true
-			break
-		}
-	}
-	if !referencesGateway {
+	if !c.referencesGateway(route) {
 		return nil
 	}
